--- main.py Changed the way agents are spawned by making them randomly generated.
--- simulation.py
New file to handle the events that take place within the simulation, like the agents spawning at the beginning of the round.
spawn_agents() creates a list of agents with random positions, directions, and speed, within set boundaries.

--- agent.py
Added max_speed and min_speed variables to maintain the agent's speed.
Added desired_speed variable to enable agent's to accelerate to a desired speed when not slowed by turning.
Added adjust_speed() function to slow the agent down when it is turning and enable it to accelerate when it is moving in a steady direction.
Added calc_wall_repulsion() function to prevent agents from sticking to walls and make them return to the swarming area when they encounter a wall.

There are several issues with the current implementation. The main issue is that agents tend to huddle together and stop accelerating altogether, instead preferring to stand still.