--- agent.py Changed speed from a class variable to be initialised when the agent is initialised.
--- movement_model.py
Added all_agents and radius as optional parameters to calculation functions to improve flexibility for testing and future changes.

calc_movement() now includes a check for zero vectors before applying normalisation to the combined vector. This is to avoid agents with non-zero speed but zero direction who get stuck in place.

--- settings_ui.py
Has beem implemented and controls the UI interactive controls for the simulation.